pyproject.toml
[tool.poetry]
name = "pytikuadapternew"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
pydantic = "^2.12.3"
aiohttp = "^3.13.2"
uvicorn = "^0.38.0"
fastapi = "==0.122.0"
sqlalchemy = "^2.0.0"
asyncpg = "^0.31.0"
python-dotenv = "^1.2.1"
greenlet = "^3.0.0"
python-jose = {extras = ["cryptography"], version = "^3.4.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
alibabacloud_dm20151123 = "1.7.2"
email-validator = "^2.3.0"
slowapi = "^0.1.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
